LawStatKG/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0

neo4j==5.23.1
python-dotenv==1.0.1

rank-bm25==0.2.2
numpy==1.26.4

sentence-transformers==3.0.1
transformers==4.44.2
torch==2.4.1